YearlyDividendServiceのrefactor
