+++ src/main/java/click/divichart/service/DividendService.java
@@ -74,6 +74,34 @@ public class DividendService {
     }
 
     /**
+     * 特定の年の配当金合計を取得するメソッドです。
+     *
+     * @param year     対象となる年
+     * @param username ユーザー名。配当のデータを取得する際に使用されます。
+     * @return 指定した年の配当金合計
+     */
+    public BigDecimal getDividendSum(int year, String username) {
+        LocalDate startDate = LocalDate.of(year, 1, 1);
+        LocalDate endDate = LocalDate.of(year, 12, 31);
+        return repository.getDividendSum(startDate, endDate, username);
+    }
+
+    /**
+     * 指定された配当増加データリストをカンマ区切りの文字列に変換します。
+     *
+     * @param dividendIncreaseData 配当増加のデータリスト。各要素は {@link BigDecimal} 型です。
+     * @return 配当増加データリストをカンマ区切りで結合した文字列。
+     * 例: "1.2,3.4,5.6"
+     */
+    public String createChartData(List<BigDecimal> dividendIncreaseData) {
+        StringJoiner chartData = new StringJoiner(",");
+        for (BigDecimal dividend : dividendIncreaseData) {
+            chartData.add(dividend.toString());
+        }
+        return chartData.toString();
+    }
+
+    /**
      * 月別配当金額を取得する
      *
      * @param targetYear 対象年
